CacheStorage: Remove disabled PutResponseWithExistingFileTest
This test attempts to test a corrupted disk condition.  Unfortunately
it is exceptionally hard to make stable since the simple disk_cache
layer is between cache_storage and the disk.  The simple disk_cache
has its own set of threads and timeout mechanisms.

This CL removes the test since the marginal benefit of testing the
condition does not merit the cost of fixing the test.  The test
was already disabled.

Bug: 1041371
Change-Id: I311392fddf8691fd27c827e582d14756f1bce89f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316692
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791474}