media: Make video encode accel tests scale by 1/2.
This CL changes the video encode accelerator test that exercises the
scaling path for zero-copy encoding so that the scaling factor is 1/2
instead of 3/4. The reason is that per b/160141652#comment3, half down
scaling is expected to be the practical case.

Note that this requires making sure that the input stream is not too
small. That's because if it's too small, scaling down might produce a
stream that's below the minimum supported resolution of the encoder.

Bug: b:160141652
Test: video.EncodeAccel.*_dmabuf on rammus and kukui.
Change-Id: I8e6b4595ac5e67bc00669f5d224c8c21a1b93edf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2315799
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791534}