Don't emit accessible text events from non-text or invisible objects
Fixes three dcheck failures, each of which is related to attempting to
emit AtkText events on objects which should not emit them. Therefore
add checks to EmitsAtkTextEvents to ensure AtkText is implemented for
the object, and that the object is not invisible or ignored.

In addition:

* For an HTMLMediaElement whose sole content is text (i.e. no player
  and no source), do not expose any children and treat as included
  but ignored.
* Always include Role::kVideo and Role::kAudio in the internal
  accessibility tree, even when ignored, for consistency.

AX-Relnotes: N/A as there is no change in end-user experience.

Bug: 1041414, 1104047, 1104048
Change-Id: Id643f148fdaba3c865c6cefe8f20151c3181135b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297431
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791541}