[cc/metrics] Fix UAF in ReportThroughputUkm
After we landed:
https://chromium-review.googlesource.com/c/chromium/src/+/2300402

There were much less crash reports than it used to be. But
we are still seeing reports.

The problem is still use-after-free. As we can see, in the
destructor of LayerTreeHostImpl, we call
    frame_trackers_.SetUkmManager(nullptr);
This sets FrameSequenceTrackerCollection::throughput_ukm_reporter_|
to nullptr. Note that FrameSequenceMetrics has a raw pointer
|throughput_ukm_reporter_| that points to the above, but that
pointer doesn't know that it is nullptr now. So in the
~FrameSequenceMetrics() which calls ReportMetrics, it tries
to use the raw pointer, which causes the problem.

Bug: 1105854, 1100868
Change-Id: I5c2045afd361a527baebce574377fbd43fd32879
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309094
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791571}