Improve caching of bounding boxes in AX renderer code
Currently in the accessibility renderer code we keep
track of a map from node ID to AXRelativeBounds
(sometimes called "location" in the code, which isn't
ideal). Every time there's a new layout, we compare the
bounds of each object to the cached bounds and send
updated bounds if it's changed.

The ultimate goal is to avoid needing to check the
bounds of every object and instead only check objects
whose bounding box likely changed. That's coming in a
follow-up.

But first a bit of refactoring will make that easier.
Rather than recreating the map from ID to bounds
every time, let's maintain a single map, and clear
node IDs when nodes are removed from the tree.

Doing that just requires a small change to
AXTreeSerializer to help keep track of node IDs
that aren't part of the tree anymore.

Bug: 1109081
Change-Id: I4c59b24645d4af74183c4b71b6a47da298c623e9
AX-Relnotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316797
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791608}