Crostini: Enable upgrade flow by default and workaround a test hang
On track to launch for realsies this milestone. Additionally, PM
guidance was that we should be enabling then merging a change to disable
(if needed) before stable rather than using Finch which is what we're
currently doing.

The test check was hanging, and since the condition was never true
before let's remove it.

Bug: None
Test: None
Change-Id: Ie16eca37622ef1ad78c655fc7999ec7a0604286f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308430
Commit-Queue: David Munro <davidmunro@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Cr-Commit-Position: refs/heads/master@{#791678}