Include printer status in ChromeVox output
To comply with GAR4 standards, this change adds the current error
printer status to the aria-description. So when the element is
navigated to in ChromeVox, the status will be announced as well.

Also fixed destination and See More label by adding single quotes. This
will pass the values into i18n as a string and not a variable.

Bug: 1059607
Change-Id: I7a5c6e715eb056534b15295d913d85e3621b9d60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2317777
Commit-Queue: Gavin Williams <gavinwill@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791733}