Add IsValidRequiredCSPAttr to network CSP parser
This CL implements the check of whether a string is a valid 'csp'
attribute according to
https://w3c.github.io/webappsec-cspee/#iframe-csp-valid-attribute-value
in the network CSP parser.

The check is still incomplete, since it relies on the Subsume
algorithm that will be implemented in future CLs.

This function will be used by the AncestorThrottle.

Bug: 1094909
Change-Id: Id325b203ba4269e542fd7b9b144da55c3e4e52d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2315683
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791763}