dpwa: Reintroduce GenerateIconLetterFromUrl fallback to generate icons.
We still need it: the title can be empty in some scenarios.

Looks like we need to generate some icon based on url
to show this icon for a user in web app install dialog, where
user should enter a non-empty title.

1) WebAppInstallTask::OnIconsRetrievedShowDialog()
2) FilterAndResizeIconsGenerateMissing()
3) Show dialog with the generated icon, `title` is still empty here.

This is a partial revert of
https://chromium-review.googlesource.com/c/chromium/src/+/2312056

The title populating code is left there in unit tests.

Bug: 1109540, 1103570
Change-Id: Ia281397d3483ad08d02ae4a9f0e2ca60bd34e859
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2319493
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791776}