Fix Password Sync in Local Sync mode (i.e. roaming profiles)
When PasswordModelTypeController was updated to support the
account-scoped password storage, a condition was added to its
GetPreconditionState() method to make sure Password Sync only runs if
either Sync-the-feature is enabled, or the user has opted in to the
account-scoped storage.

However, for the case of local Sync, Sync-the-feature is *not*
considered enabled (and of course no account-storage-opt-in exists).
So this broke local Sync for Passwords.

Bug: 1108609
Change-Id: I2bb2edb7247d3b2e95dba306718488352329840b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316214
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791782}