Add RemoteApps
This CL implements the RemoteApps PublisherBase and RemoteAppsManager
which are used to manage Remote Apps. Remote Apps are items which are
added to the launcher but are not backed by an app/extension/web app.
Instead, the item is managed by an extension via an API. The extension
is able to add multiple items and folders to the launcher and observe
when the items are launched. The extension can then decide on how to
handle the launch event. Currently RemoteApps are only available to
Managed Guest Sessions.

RemoteAppsManager is the KeyedService which handles the adding and
deleting of Remote Apps and folders. RemoteAppsManager also implements
RemoteApps::Delegate and is used by RemoteApps to handle the LaunchApp(),
GetIcon() and GetMenuModel() methods.

RemoteAppsManager is not initialized as a KeyedService so this CL does
not change any existing behavior.

Bug: 1101208
Change-Id: I664469fa0284c6e65288e59ba7a5d1271b081a04
DD: go/cros-remote-apps
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2299226
Commit-Queue: Jit Yao Yap <jityao@google.com>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791788}