crostini: Fix the penguin fallback icon not being loaded
Crostini registers Linux applications it knows about, however many GUI
Linux apps don't have .desktop files to register or icons to show but
they still have windows that should appear in the shelf. For them we
show a default penguin icon as a fallback when we can't find a better
icon. This was broken per the linked bug.

There are two changes to fix this:
1. We end up in RequestContainerAppIcon, however this method ignores
requests for unregistered apps and never calls the callback. Fix that by
skipping directly to returning no icon instead of calling
RequestContainerAppIcon
2. Pass app_id to LoadIcon in crostini_apps so it can tell if it's being
asked for an icon for an unregistered Linux app (as opposed to e.g. an
unregistered PluginVM app) and update a check to handle this
appropriately.

Bug: chromium:1106280
Test: manual - run some Crostini apps and look for penguins

Change-Id: Iacf7c889989a4e91db16adf1244f13e56a71e97b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318891
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: David Munro <davidmunro@google.com>
Cr-Commit-Position: refs/heads/master@{#791792}