Remove use of setTimeout in ref test for web-animations
Use requestAnimationFrames instead of setTimeouts in the ref tests to
accommodate variability in timing. The background color is animated
between green and red to better align with reftest guidelines.

The test to slow down the animation was removed since it would be
essentially equivalent to the zero playback rate test after the
refactoring.

Bug: 1108342, 1064127
Change-Id: Ie933b68cf45caabfa70fbe5deb881e120ea5e28a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2313157
Reviewed-by: Xida Chen <xidachen@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791833}