[Extensions] Add ParallelUnpacker class
This class makes it easy to starts multiple SandboxedUnpacker's in
parallel, and get notified as soon as any of them completes. Since
SandboxedUnpacker is disk-bound and often takes >2s to run, running them
in parallel can be beneficial.

Right now ParallelUnpacker is not used anywhere, but a follow-up CL will
use it  to make extensions update in parallel, which will speed up
extension installs when many of them are queued at the same time (such
as for force-installed extensions).

Bug: 1103447
Change-Id: Ibc35588134fe0e91c4f7a243677181a08c642552
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2294563
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791835}