[css-lists] Fix marker not correctly updated in legacy layout
If a list item contains a block box, the layout tree is initially like:

  LayoutListItem {LI}
    LayoutBlockFlow (anonymous)
      LayoutListMarker {::marker}
    LayoutBlockFlow {P}
      LayoutText #text

However, an outside ::marker should appear in the same line as the first
non-marker text of the list item. Therefore, it's changed into:

  LayoutListItem {LI}
    LayoutBlockFlow (anonymous)
    LayoutBlockFlow {P}
      LayoutListMarker {::marker}
      LayoutText {#text}

But this was not happening when dynamically toggling list-style-type
between 'none' and some other value. The reason was that there was a
'NormalChildNeedsLayout()' condition, which would hold the 1st time, but
in following layouts it would be false.

This patch just removes that condition.

Bug: 1107783

TEST=external/wpt/css/css-lists/change-list-style-type-002.html

Change-Id: Ic82f1a474b0a0ddf2f2aab8a60afac299aa6e2d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310359
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/master@{#791857}