Inject a CrxDownloaderFactory in UpdateClient.
This CL defines a CrxDownloaderFactory interface, for which
UpdateClient provides an implementation. This interface is injected
using a function of Configurator. Callers of Update client such as
the component and extensions updater use the Chromium
implementation of the factory: there is no change in the existing
behavior.

Projects such as chrome/updater could in the future provide their
own implementation of the factory to override the CrxDownloader stack.

Bug: 1109283

Change-Id: I36f973b470a756f7e428bdac0407986374488c43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302716
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791972}