Add privacy budget UKM collection for local font lookups
Expands the FontMatchingMetrics object to record (successful or
unsuccessful) local font lookup events in a range of situations, as
enumerated in FontLookupType. The font lookup inputs (e.g. font name)
and a hash of the resulting font data are periodically published.
This will allow for measurement of the privacy impact of local font use.

This cl does not record any font lookup events associated with a binary
or remote CSSFontFaceSource. Metrics reporting in the
OffscreenFontSelector will occur in a separate cl (crrev.com/c/2197604).

Bug: 973801
Change-Id: I8304568af91fc56c1efe4087a9a7effdad234474
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2195493
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791976}