[Fuchsia] Wait for CDM OnProvisioned event
The fuchsia.media.drm.ContentDecryptionModule sends an event upon the
completion of provisioning[1]. This is an indication that the CDM is
ready for use. This CL pends the CDM creation callback until after the
OnProvisioned event has been received, ensuring the CDM is ready prior
to any calls being made to it.

[1] https://fuchsia.dev/reference/fidl/fuchsia.media.drm#fuchsia.media.drm/ContentDecryptionModule.OnProvisioned

Bug: fuchsia: 13619
Change-Id: Ie119daa58610b2cab0cade14b7dbf1ef93b3dcfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2315000
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791987}