[AppBanner] Retry requests to InstallableManager on manifest url change.
This change adds MANIFEST_URL_CHANGED to InstallableStatusCodes, and
adds support for InstallableManager reporting this error (as well as
reporting when a navigation happens or when the renderer exists).

Then it modifies the AppBannerManager class to listen for this result
when it fetches the manifest or checks if a tab is installable.

Finally, it modifies AppBannerManager to also listen for manifest url
changes, and when the occur outside of calls to the InstallableManager
(as these will be reported by the method above), it resets its state
and restarts the request.

Possible upgrades:
 * Add the rate limiter in AppBannerManager for retrying (or just count
   the number of retries & stop after X times)

Design doc:
https://docs.google.com/document/d/1HjN4JL9DCaTNuLj29ieqJEsexpfAzfJnoVrqY7SWvd8/edit

R=dominickn@chromium.org

Bug: 792299
Change-Id: Ia26c727d68d50ebee1966ed0e0318874e1843fe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300828
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Auto-Submit: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792009}