Fix naming issue for matrix compound test suites that have the 'test' key
If the 'test' key is in the definition for a basic test suite, then
that value overrides the mixed compound test suites name.
The name for the suite should be the test suites name plus the
identifier field set in a variants list entry.

Bug: 1041619
Bug: 1109967
Change-Id: I8a44b23d2511d2a5e85c0d06e3047c1d50abf011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318305
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792036}