viz: Refactor SetEnableDCLayers logic out of DirectRenderer
Refactor the SetEnableDCLayers calls out of DirectRenderer into
OverlayProcessorWin.  Having the logic in the overlay processor is
needed for deferring overlay promotion to subsequent frames when partial
quad aggregation is enabled and full damage wasn't aggregated.

As a result move the plumbing of SetEnableDCLayers from DirectRenderer
to OutputSurface since the former isn't meant to be used by the overlay
processor.

Bug: 1077210
Change-Id: I191ba9ea591084d8f054d56958d6056ce1507a57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302742
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Reviewed-by: weiliangc <weiliangc@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792045}