[Nearby] Add nearby-device-icon component
This renders a device's icon. For now it only looks at the device type
but in future this will show an image if the remote device has one
defined. Used in the nearby-device component on the discovery page and
in the nearby-progress component on the confirmation page.

Bug: 1103190
Change-Id: I9e02048706406ee174c71c869af15938888c5bb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2315688
Commit-Queue: Richard Knoll <knollr@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792064}