Ignore other payment methods when app store billing is available in TWA
This patch adds a IsPreferred bit to a PaymentApp, which indicates that
a payment app should be chosen by default and others are to be ignored.
If an app store billing app is found and the payment request is within
a TWA, then the app store billing app is set as preferred.

Note that the browser test in
payment_request_app_store_billing_browsertest.cc currently just tests
that requesting the play billing payment method inside of a TWA throws a
generic NotSupportedError, instead of the play billing specific error
that is thrown when outside of a TWA. We'll need to first mock some more
pieces of AndroidPaymentAppFinder first to truly test this behavior.

Bug: 1095827
Change-Id: Ib91b1e03b89c800abaa51d25ff699eea9d1b2914
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254058
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792084}