indexeddb: reset async tasks in ~WebIDBGetDBNamesCallbacksImpl
Since sometimes the WebIDBGetDBNamesCallbacksImpl can be destroyed when
the promise is resolved, make sure that no code that could reference it
is still around.  Store the async task as an optional member so it can
be cleared during the destructor.

Followup to:
https://chromium-review.googlesource.com/c/chromium/src/+/2311620

Bug: 1109467
Change-Id: Id6a0ff0a3703fab94e9684e41f16d5a1bac20468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2321332
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Auto-Submit: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792121}