SetNeedsPaint on tables when collapsed borders change
Tables always paint collapsed borders, even if the rows are seprately
composited. While we don't need to invalidate paint for the table
(because the row is the DisplayItemClient for the borders), we do need
to cause repaint for its painting layer, because the table paints those
borders.

See also (*), which removed the code that previously had the same
effect. There was previously no test for this code.

(*) https://chromium-review.googlesource.com/c/chromium/src/+/2317848

Change-Id: I269101d3bf28445c0ec3d1f6bcad1a23cf3e46d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2321234
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792138}