[bookmarks] Fix missing save-to-disk if shut down during backup
If the browser is shut down while in BACKUP_DISPATCHED (i.e. the task
was posted but didn't complete yet), then prior to this patch there was
a risk that BookmarkStorage was destroyed before the intended save took
place, because there was no scheduled write in ImportantFileWriter.

This situation should be rare in practice because making a backup is
fast, but does lead to test flakiness, and can be reliably reproduced
(e.g. with BookmarkBrowsertest.EmitUmaForDuplicates) if an artificial
delay is introduced when generating the backup file.

While fixing the issue, a few implementation details have been
simplified:
1) Backup-generation does not require waiting until it completes before
   scheduling a regular write, since a SequencedTaskRunner guarantees
   the ordering in which the tasks will run (and copying a file is a
   synchronous operation).

2) There is no need to manually implement SaveNow(), since
   ImportantFileWriter does the same in DoScheduledWrite().

Change-Id: Ib1de11935b4e37d44e4ad79eb77365cfefb4ff7b
Bug: 1108803
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320869
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792168}