[FlexNG] Fix flex-item grid layout explosion.
When we had a column flex-item, which was also a grid, where the percent
template added up to something >100% on each layout the grid would grow
larger.

This was caused by some logic to fix relative-positioned children in:
https://chromium-review.googlesource.com/c/chromium/src/+/2093273

However when this value was read within layout, it could refer to the
previous layout. If numerous layouts occurred sequentially, this would
cause the layouts to grow unbounded.

This patch disables this path when read during layout for a flex-item
(in NG). During layout the flex-item should have a BoxLayoutExtraInput
set.

Bug: 1108928
Change-Id: I4d0d41755c9d33eb2b4c217ee6bfdb1acd48cbbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320418
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792193}