Prevent permission prompts on behalf of inactive RenderFrameHosts
Make sure we never show a prompt on behalf of a page that is not current
for example because it was moved to the back forward cache. For that we
need to associate all requests with the originating frame
(RenderFrameHost). Before we actually show the prompt we check whether
the frame still exists (could have been deleted) and is current. If that
is not the case we just cancel the request.

This is an extra safeguard on top of what is already done in
PermissionRequestManager::DidFinishNavigation, which only clears queued
requests at the time of the navigation commit, but does not prevent us
from accepting requests after that. Ideally we would not cancel those
queued requests but store them on the side in case the frame becomes
current again (gets out of the cache).

Change-Id: Ib7eaa6b935b96cfe76d9f8cbcd7a3712900855fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2307212
Commit-Queue: Carlos Caballero <carlscab@google.com>
Reviewed-by: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792204}