COOP: Check opener exist in GetSourceRfhForCoopReporting()
If we don't, then a crash will happens when whenever the user tries to
restore a tab on a URL using Cross-Origin-Opener-Policy-Report-Only
(COOPRO). A test depending on this patch will be added on this branch.

The current code is still potentially incorrect. Some TODO are added.

Bug: chromium:922191
Change-Id: I10c2e08daf3aff3b3de7ee6c0eee90b1f3e777c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2317964
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Arthur Hemery <ahemery@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792212}