Set CFBundleName for SxS Chromes
In the past, this wasn't done for Canary because the idea was for
it to be as identical to the normal Chrome Stable as possible. Now
that we're making Chromes to run side-by-side to Chrome Stable,
they need their own identity lest visual issues like the one in
the linked bug appear.

Bug: 1109411
Change-Id: Ie3c75014f5dabfe61d5925383dac5f07b2fce481
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318726
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Auto-Submit: Avi Drissman <avi@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792241}