Fix taskbar context menu icon for UserDataDir policy.
Back in 2012, we had a problem with the state of the created
shortcuts on the desktop: crbug.com/133585

As a fix, we create full-fledged shortcuts and icons only if
the profile with which the browser is launched is in the
DefaultUserFolder.
If the UserDataDir policy is used, a shortcut is created in
"%AppData%\Roaming\Microsoft\Internet Explorer\Quick Launch\
User Pinned\ImplicitAppShortcuts" with the path to the icon
%ProfilePath%\Default\Yandex Profile.ico, but the icon itself
is not created in the profile.

To fix this problem, we need to subscribe to create shortcuts
and enable the feature of creating shortcuts for a profile
from a policy, if such exists.

When using --user-data-dir, a different mechanism is used,
in this case the shortcut is not created, but the crhome.exe
icon is used.

BUG=1108485

Change-Id: I21d914d5a4d8b18960bda248a90bf320b649e323
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2314516
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792243}