[Switch access] Don't navigate to occluded windows.
Check if the top center of a window is visible before moving
SA focus to that window.

It's still possible to end up at an occluded window when restoring
SA focus from history, or by going to firstValidChild. We could
add an eventhandler to the desktop or a repeated interval to
each window node child to check for occlusion. This is left as
a TODO.

AX-Relnotes: N/A
Bug: 1106080
Change-Id: Ib88d53a3b18b8c9f5da2403b61d1624e1258014c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318245
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#792318}