Refactor base::Callback and base::Bind in chrome/browser/media/webrtc
This CL converts instances of base::Callback and base::Bind into
base::OnceCallback/base::RepeatingCallback and base::BindOnce/
base::BindRepeating, as part of the Chrome code health project.

In this case all of the `GenericDoneCallback` etc types were effectively
OnceCallbacks, so this CL replaces them with that, as they're generally
more efficient and provide more flexibility to the creator of the
callback.

Bug: 1007635
Change-Id: I63b70bc3e46497d979d5f524a533621ad824c0d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318004
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Commit-Position: refs/heads/master@{#792374}