[CryptAuth] Add IsListening() method to GCM manager
The method returns true if the CryptAuth app handler has been added to
the GCM driver. This is particularly important to know when fetching
Instance ID data (see crbug.com/1006280). Child CLs will utilize this
method to confirm that the manager is listening before registering with
GCM or fetching Instance IDs.

---

This is ClientAppMetadata change-detection CL [1/4]:

This chain of CLs relates to the end goal of re-enrolling if
ClientAppMetadata has changed since the last successful enrollment. For
example, when Phone Hub become supported, we will want to re-enroll.

We were previously only fetching ClientAppMetadata on demand, for
instance, right before a periodic enrollment. Now, we need to fetch the
data every time the service starts up to see if it changed. We continue
to make the assumption that the data will not change during a user
session.

Bug: 1105703
Change-Id: Ifb08bc479b96e0ace0f85d27041a8259b9b29967
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2319327
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Josh Nohle <nohle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792408}