[PM] Cleanup the site data writer logic
This cleanup some things inherited from the previous implementation,
some properties were tracked multiple time (in the PageNode itself, in
the SiteDataWriter and in the NodeData object), this removes some of the
duplication and privileges using data directly from the PageNode.

This also renames the NotifySiteVisibilityChanged function to a pair of
NotifySiteForegrounded/NotifySiteBackgrounded function for consistency
with the other functions of the SiteDataWriter class.

Change-Id: I6412e2f79d8f330214567a439c33727e87210ce0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283297
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792438}