Transform the selection rect in vertical writing mode
In InlineTextBoxPainter and NGTextFragmentPainter, previously we
manually transpose the selection rect to convert it between the space of
the original space and the rotated space for vertical text. The code
might be inconsistent with the rotation transform. Actually
crrev.com/c/2319540 exposed the bug in InlineTextBoxPainter when the
vertical selection rect is wider than the text rect.

Now use the rotation transform to transform the selection rect to avoid
the inconsistency.

Change-Id: If75fa61539842cc32a9129391b1a32edbe631847
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324066
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792443}