Use Views context menus on tabstrip on Mac.
This was the result of a discussion with ellyjones@ regarding the best
way to handle features that could only easily be implemented in Views
that we needed for user education.

The decision was made that context menus in Top Chrome could be moved to
Views, but not for web contents and context menus on text fields, which
would remain in Cocoa to provide visual distinction and to provide
access to e.g. the system emoji chooser.

This CL adds a mechanism by which Top Chrome code can force a context
menu that would otherwise be rendered natively to use Views code
instead, and applies it to tabstrip context menus.

Bug: 1109256
Change-Id: Ib3c3970f56df0714687aaf87a1267d7ae8e2c17b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2317809
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792451}