Speculative fix for a bug in ProfileDestroyer destruction
This CL tentatively reverts following 6 CL's as a speculative fix
for an ANR issue around ProfileDestroyer:

 https://crrev.com/c/2171463
  "Clean up ProfileDestroyer."
 https://crrev.com/c/2134019
  "Add tracing to make it easier to debug shutdown behavior."
 https://crrev.com/c/2113391
  "Android: Handle multiple ProfileDestroyers correctly"
 https://crrev.com/c/2032385
  "Fix a crash in ProfileDestroyer due to delayed RPH destruction"
 https://crrev.com/c/2022791
  Revert "Only CHECK for RPHs that outlive an OTR profile in release"
 https://crrev.com/c/1971031
  "Android: Fix a bug when closing all incognito tabs for preview tab"

The first CL(1071031) landed to fix Issue 1029677, and subsequent CL's
handled the remaining/new issues reported after that.

The ANR bug (1095078/1095548) is suspected to have to do with
ProfileDestroyer destruction of incognito profile. Not having found
good clues, this is trying to see if the series of these changes
in the class caused it. With this, the state of ProfileDestroyer
is restored back to M80 which had no issues.

This is NOT a permanent solution for the problem.  Will reland
the above CL's if they are proven innocent, or find a right fix
if they indeed are the culprit.

TBR=msarda@chromium.org

(cherry picked from commit ba2b87b8507be97c3ab52aa656c898d4ae6fe467)

Bug: 1029677, 1040839, 1095078, 1095548
Change-Id: I761809a90cd4f3a866ae7754da4dab34640dcb3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308616
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#791182}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324684
Cr-Commit-Position: refs/branch-heads/4183@{#1019}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}