[CrOS Settings] Reverse scrolling touchpad change.
Prior to this change, enabling Reverse scrolling would meant that
swiping up would cause the page to move up, and vice versa. This CL
corrects that so that:

* Turn off Reverse (swipe up to move the page up).
* Turn on Reverse (swipe up to move the page down).

This is fixed by setting the "natural_scroll" pref to true if
"Reverse Scrolling" is enabled, and the "natural_scroll" pref to false
if "Reverse Scrolling" is disabled, instead of the opposite as it was
previously.

Fixed: 1109966
Change-Id: Id60df73f421992dce893339b6efaac0941ef9110
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2321235
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#792558}